requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0 ## git+https://github.com/ageitgey/face_recognition_models
idna==3.10
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
